rag-prt/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
atlassian-python-api==4.0.7
attrs==25.4.0
beautifulsoup4==4.14.3
boto3==1.42.9
botocore==1.42.9
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
dataclasses-json==0.6.7
deprecated==1.3.1
docx2txt==0.9
faiss-cpu==1.13.1
frozenlist==1.8.0
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.1.2
langchain-aws==1.1.0
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.1.1
langchain-text-splitters==1.0.0
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.0
langsmith==0.4.59
lxml==6.0.2
marshmallow==3.26.1
multidict==6.7.0
mypy-extensions==1.1.0
numpy==2.3.5
oauthlib==3.3.1
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
pdfminer-six==20251107
propcache==0.4.1
pycparser==2.23
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pyyaml==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
s3transfer==0.16.0
six==1.17.0
soupsieve==2.8
sqlalchemy==2.0.45
tenacity==9.1.2
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
urllib3==2.6.2
uuid-utils==0.12.0
wrapt==2.0.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0